update tokenize function
